requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# resy-bot/notify.py: 1
discord_webhook == 0.13.0

# resy-bot/bot.py: 12
python_dotenv == 0.15.0

# resy-bot/bot.py: 5
pytz == 2020.5

# resy-bot/bot.py: 6,7,8,9,10,11
selenium == 3.141.0
